Repository

+++ CinemaApp.Web/Controllers/MovieController.cs
@@ -1,35 +1,36 @@
-﻿using CinemaApp.DataAccess.Data;
+﻿using CinemaApp.DataAccess.Repository.IRepository;
 using CinemaApp.Models.DomainModels;
 using CinemaApp.Models.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using Utility;
 
 namespace CinemaApp.Web.Controllers
 {
-    public class MovieController : Controller
+	public class MovieController : Controller
 	{
-		private readonly ApplicationDbContext _db;
+		private readonly IUnitOfWork _unitOfWork;
 		private readonly IWebHostEnvironment _webHostEnvironment;
 
-		public MovieController(ApplicationDbContext db, IWebHostEnvironment webHostEnvironment)
+		public MovieController(IUnitOfWork unitOfWork, IWebHostEnvironment webHostEnvironment)
 		{
-			_db = db;
+            _unitOfWork = unitOfWork;
 			_webHostEnvironment = webHostEnvironment;
 		}
 
-        [Authorize(Roles = SD.Role_Employee)]
-        public IActionResult Index()
-		{	
-			var movies = _db.Movies.Include(x => x.MovieGenres).ThenInclude(x => x.Genre).AsNoTracking().ToList();
+		//[Authorize(Roles = SD.Role_Employee)]
+		public async Task<IActionResult> Index()
+		{
+			var movieGenres = await _unitOfWork.MovieGenre.GetAllAsync(includeProperties: "Movie,Genre");
+			var movies = movieGenres.Select(x => x.Movie).Distinct().ToList();
+
 			return View(movies);
 		}
 
-        [Authorize(Roles = SD.Role_Employee)]
-        public IActionResult Create()
-		{	
-			var genres = _db.Genres.ToList();
+		//[Authorize(Roles = SD.Role_Employee)]
+		public async Task<IActionResult> Create()
+		{
+			var genres = await _unitOfWork.Genre.GetAllAsync();
 			MovieVM movieVM = new MovieVM
 			{
 				Genres = genres,
@@ -39,16 +40,16 @@ namespace CinemaApp.Web.Controllers
 			return View(movieVM);
 		}
 
-        [Authorize(Roles = SD.Role_Employee)]
-        [HttpPost]
-		public IActionResult Create(MovieVM movieVM, IFormFile? formFile)
+		//[Authorize(Roles = SD.Role_Employee)]
+		[HttpPost]
+		public async Task<IActionResult> Create(MovieVM movieVM, IFormFile? formFile)
 		{
 			if (ModelState.IsValid)
-			{	
+			{
 				if (movieVM.SelectedGenreIds.Count <= 0)
 				{
 					ModelState.AddModelError("Movie.MovieGenres", "Movie must belongs to at least one genre.");
-					movieVM.Genres = _db.Genres.ToList();
+					movieVM.Genres = await _unitOfWork.Genre.GetAllAsync();
 					return View(movieVM);
 				}
 
@@ -63,55 +64,65 @@ namespace CinemaApp.Web.Controllers
 						formFile.CopyTo(fileStream);
 					}
 					movieVM.Movie.ImageUrl = @"\images\movie\" + fileName;
+				} else
+				{
+					movieVM.Movie.ImageUrl = @"\images\default\default-movie-poster.jpg";
 				}
 
 				movieVM.Movie.MovieGenres = movieVM.SelectedGenreIds.Select(id => new MovieGenre { GenreId = id }).ToList();
 
-				_db.Movies.Add(movieVM.Movie);
-				_db.SaveChanges();	
+				await _unitOfWork.Movie.AddAsync(movieVM.Movie);
+				await _unitOfWork.SaveAsync();
 
 				return RedirectToAction("Index");
 			}
 			return View(movieVM);
 		}
 
-		public IActionResult Details(int? id)
+		public async Task<IActionResult> Details(int? id)
 		{
-			var movie = _db.Movies.Where(x => x.Id == id).Include(x => x.MovieGenres).ThenInclude(x => x.Genre).First();
+			var movieGenres = await _unitOfWork.MovieGenre.FindAsync(x => x.MovieId == id, includeProperties: "Movie,Genre");
+			var movie = movieGenres.Select(x => x.Movie).FirstOrDefault();
+
 			return View(movie);
 		}
 
-        [Authorize(Roles = SD.Role_Employee)]
-        public IActionResult Edit(int? id)
+		//[Authorize(Roles = SD.Role_Employee)]
+		public async Task<IActionResult> Edit(int? id)
 		{
 			if (id == null || id == 0)
+			{
 				return NotFound();
+			}
 
-			var movie = _db.Movies.Where(x => x.Id == id).Include(x => x.MovieGenres).ThenInclude(x => x.Genre).First();
+			var movieGenres = await _unitOfWork.MovieGenre.FindAsync(x => x.MovieId == id, includeProperties: "Movie,Genre");
+			var movie = movieGenres.Select(x => x.Movie).FirstOrDefault();
 
-			if (movie == null) 
+			if (movie == null)
+			{
 				return NotFound();
-			
+			}
+
 			MovieVM movieVM = new MovieVM
 			{
 				Movie = movie,
 				SelectedGenreIds = movie.MovieGenres.Select(x => x.GenreId).ToList(),
-				Genres = _db.Genres.ToList()
+				Genres = await _unitOfWork.Genre.GetAllAsync()
 			};
 
 			return View(movieVM);
 		}
 
-        [Authorize(Roles = SD.Role_Employee)]
-        [HttpPost]
-		public IActionResult Edit(MovieVM movieVM, IFormFile? formFile)
+		//[Authorize(Roles = SD.Role_Employee)]
+		[HttpPost]
+		public async Task<IActionResult> Edit(MovieVM movieVM, IFormFile? formFile)
 		{
 			if (ModelState.IsValid)
 			{
 				if (movieVM.SelectedGenreIds.Count <= 0)
 				{
 					ModelState.AddModelError("Movie.MovieGenres", "Movie must belongs to at least one genre.");
-					movieVM.Genres = _db.Genres.ToList();
+					movieVM.Genres = await _unitOfWork.Genre.GetAllAsync();
 					return View(movieVM);
 				}
 
@@ -125,7 +136,7 @@ namespace CinemaApp.Web.Controllers
 					{
 						var oldImagePath = Path.Combine(wwwRootPath, movieVM.Movie.ImageUrl.TrimStart('\\'));
 
-						if (System.IO.File.Exists(oldImagePath))
+						if (System.IO.File.Exists(oldImagePath) && !oldImagePath.Contains("default"))
 						{
 							System.IO.File.Delete(oldImagePath);
 						}
@@ -138,7 +149,7 @@ namespace CinemaApp.Web.Controllers
 					movieVM.Movie.ImageUrl = @"\images\movie\" + fileName;
 				}
 
-				movieVM.Movie.MovieGenres = _db.MovieGenres.Where(x => x.MovieId == movieVM.Movie.Id).ToList();
+				movieVM.Movie.MovieGenres = (ICollection<MovieGenre>) await _unitOfWork.MovieGenre.FindAsync(x => x.MovieId == movieVM.Movie.Id);
 
 				foreach (var genreId in movieVM.SelectedGenreIds)
 				{
@@ -152,38 +163,45 @@ namespace CinemaApp.Web.Controllers
 				{
 					if (!movieVM.SelectedGenreIds.Contains(movieGenre.GenreId))
 					{
-						_db.MovieGenres.Remove(movieGenre);
+						_unitOfWork.MovieGenre.Remove(movieGenre);
 					}
 				}
 
-				_db.Movies.Update(movieVM.Movie);
-				_db.SaveChanges();
+				_unitOfWork.Movie.Update(movieVM.Movie);
+				await _unitOfWork.SaveAsync();
+
 				return RedirectToAction("Index");
 			}
 			return View();
 		}
 
-        [Authorize(Roles = SD.Role_Employee)]
-        public IActionResult Delete(int? id)
+		//[Authorize(Roles = SD.Role_Employee)]
+		public async Task<IActionResult> Delete(int? id)
 		{
 			if (id == null || id == 0)
-				return NotFound();
-
-			var movie = _db.Movies.Find(id);
-
-			if (movie == null)
-				return NotFound();
-
-			if(!string.IsNullOrEmpty(movie.ImageUrl))
 			{
-				var fileToDeletePath = Path.Combine(_webHostEnvironment.WebRootPath, movie.ImageUrl.TrimStart('\\'));
-
-				if (System.IO.File.Exists(fileToDeletePath))
-					System.IO.File.Delete(fileToDeletePath);
+				return NotFound();
 			}
 
-			_db.Movies.Remove(movie);
-			_db.SaveChanges();
+			var movie = await _unitOfWork.Movie.GetAsync(id);
+
+			if (movie == null)
+			{
+				return NotFound();
+			}
+
+			if (!string.IsNullOrEmpty(movie.ImageUrl))
+			{
+				var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, movie.ImageUrl.TrimStart('\\'));
+
+				if (System.IO.File.Exists(oldImagePath) && !oldImagePath.Contains("default"))
+				{
+					System.IO.File.Delete(oldImagePath);
+				}
+			}
+
+			_unitOfWork.Movie.Remove(movie);
+			await _unitOfWork.SaveAsync();
 
 			return RedirectToAction("Index");
 		}
